bmed-lme-app/requirements.txt
requests
pandas
beautifulsoup4
reportlab
pypdf
flask
python-dotenv